Move the animated heading to its own component
