+++ client/src/main/scripts/components/homepage/WhatIsThis.tsx
@@ -17,7 +17,8 @@
  *
  * You can contact us at team@mytake.org
  */
-import React, { useEffect, useState } from "react";
+import React from "react";
+import AnimatedHeading from "./AnimatedHeading";
 import BlinkingCursor from "./BlinkingCursor";
 import HomeSection from "./HomeSection";
 
@@ -26,83 +27,11 @@ interface WhatIsThisProps {
   rightSocial: React.ReactElement;
 }
 
-interface HeadingContent {
-  start: string;
-  endings: string[];
-}
-const HEADING_TEXT: HeadingContent = {
-  start: "Is this ",
-  endings: ["out of context?", "what was said?"],
-};
-const TYPING_DELAY = 100;
-const STATIC_DELAY = 2000;
-const DELETING_DELAY = 10;
-type TypingState = "TYPING" | "STATIC" | "DELETING";
 const WhatIsThis: React.FC<WhatIsThisProps> = (props) => {
-  const [headingText, setHeadingText] = useState(HEADING_TEXT.start);
-  const [endingIndex, setEndingIndex] = useState(0);
-  const [typingState, setTypingState] = useState<TypingState>("TYPING");
-
-  const continueTyping = (currentEndingIndex: number) => {
-    setTimeout(() => {
-      setHeadingText(
-        (prevText) =>
-          HEADING_TEXT.start +
-          HEADING_TEXT.endings[currentEndingIndex].substr(
-            0,
-            prevText.length + 1 - HEADING_TEXT.start.length
-          )
-      );
-    }, TYPING_DELAY);
-  };
-
-  const continueDeleting = (currentEndingIndex: number) => {
-    setTimeout(() => {
-      setHeadingText(
-        (prevText) =>
-          HEADING_TEXT.start +
-          HEADING_TEXT.endings[currentEndingIndex].substr(
-            0,
-            prevText.length - 1 - HEADING_TEXT.start.length
-          )
-      );
-    }, DELETING_DELAY);
-  };
-
-  useEffect(() => {
-    switch (typingState) {
-      case "TYPING":
-        if (
-          headingText !==
-          HEADING_TEXT.start + HEADING_TEXT.endings[endingIndex]
-        ) {
-          continueTyping(endingIndex);
-        } else {
-          setTypingState("STATIC");
-        }
-        break;
-      case "STATIC":
-        setTimeout(() => {
-          setTypingState("DELETING");
-        }, STATIC_DELAY);
-        break;
-      case "DELETING":
-        if (headingText !== HEADING_TEXT.start) {
-          continueDeleting(endingIndex);
-        } else {
-          setTypingState("TYPING");
-          setEndingIndex((prevIndex) =>
-            prevIndex >= HEADING_TEXT.endings.length - 1 ? 0 : prevIndex + 1
-          );
-        }
-        break;
-    }
-  }, [endingIndex, headingText, typingState]);
-
   return (
     <HomeSection>
       <h2 className="home__h1 home__h1--animated">
-        Are we lying to you? <span>{headingText}</span>
+        Are we lying to you? <AnimatedHeading />
         <BlinkingCursor />
       </h2>
       <p className="home__body home__body--center">Click one to see</p>
